+++ backend/providers/src/test/kotlin/ru/sokomishalov/memeory/providers/AbstractProviderIntegrationTest.kt
@@ -4,13 +4,7 @@ package ru.sokomishalov.memeory.providers
 
 import kotlinx.coroutines.runBlocking
 import org.junit.Assert.*
-import org.junit.Before
 import org.junit.Test
-import org.junit.runner.RunWith
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc
-import org.springframework.boot.test.context.SpringBootTest
-import org.springframework.test.context.junit4.SpringRunner
 import ru.sokomishalov.commons.core.log.Loggable
 import ru.sokomishalov.commons.core.serialization.OBJECT_MAPPER
 import ru.sokomishalov.memeory.core.dto.ChannelDTO
@@ -21,27 +15,17 @@ import ru.sokomishalov.memeory.core.util.image.getImageByteArray
 /**
  * @author sokomishalov
  */
-@RunWith(SpringRunner::class)
-@SpringBootTest
-@AutoConfigureMockMvc
-abstract class AbstractProviderIntegrationTest : Loggable {
+abstract class AbstractProviderIntegrationTest {
+
+    companion object : Loggable
 
     protected abstract val channel: ChannelDTO
-
-    private lateinit var service: ProviderService
-
-    @Autowired
-    lateinit var providerFactory: ProviderFactory
-
-    @Before
-    fun setUp() {
-        service = providerFactory[channel.provider]!!
-    }
+    protected abstract val service: ProviderService
 
     @Test
     fun `Check that channel memes has been fetched`() {
         val memes = runBlocking { service.fetchMemes(channel) }
-        log("Memes: ${OBJECT_MAPPER.writerWithDefaultPrettyPrinter().writeValueAsString(memes)}")
+        logInfo("Memes: ${OBJECT_MAPPER.writerWithDefaultPrettyPrinter().writeValueAsString(memes)}")
 
         assertFalse(memes.isNullOrEmpty())
         memes.forEach {
